Update the code to work with the component version of the domain contact details form
